+++ src/Symfony/Component/Mailer/Bridge/Amazon/Tests/Transport/SesHttpAsyncAwsTransportTest.php
@@ -112,7 +112,7 @@ class SesHttpAsyncAwsTransportTest extends TestCase
 
         $message = $transport->send($mail);
 
-        $this->assertSame('foobar', $message->getInfo('id'));
+        $this->assertSame('foobar', $message->getMetadata('id'));
         $this->assertNotSame('foobar', $message->getMessageId());
     }