+++ src/Symfony/Component/Mailer/CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 8.1
 ---
 
- * Add `SentMessage::addInfo()`, `SentMessage::getInfo()`, and `SentMessage::getAllInfo()` to allow transports to store transport-specific metadata on a sent message
+ * Add `SentMessage::addMetadata()`, `SentMessage::getMetadata()`, and `SentMessage::getAllMetadata()` to allow transports to store transport-specific metadata on a sent message
 
 8.0
 ---